Renamed filenames to reflect class names
